pyproject.toml
[project]
name = "b1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "typer>=0.24.1",
]

[project.scripts]
b1 = "b1.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
